requirements-dev.txt
-r requirements.txt

# ── Testing ────────────────────────────────────
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1

# ── Linting & Formatting ──────────────────────
ruff==0.9.4

# ── Debug ──────────────────────────────────────
django-debug-toolbar==5.0.1
ipython==8.31.0
